moved TasksWidget to publisher
