+++ client/ayon_core/tools/publisher/widgets/tasks_widget.py
@@ -1,7 +1,11 @@
-from qtpy import QtCore, QtGui
+from qtpy import QtWidgets, QtCore, QtGui
 
-from ayon_core.tools.utils.tasks_widget import TasksWidget, TASK_NAME_ROLE
+from ayon_core.tools.utils.views import DeselectableTreeView
 from ayon_core.tools.utils.lib import get_default_task_icon
+
+TASK_NAME_ROLE = QtCore.Qt.UserRole + 1
+TASK_TYPE_ROLE = QtCore.Qt.UserRole + 2
+TASK_ORDER_ROLE = QtCore.Qt.UserRole + 3
 
 
 class TasksModel(QtGui.QStandardItemModel):
@@ -141,15 +145,159 @@ class TasksModel(QtGui.QStandardItemModel):
         return super(TasksModel, self).headerData(section, orientation, role)
 
 
-class CreateWidgetTasksWidget(TasksWidget):
+class TasksProxyModel(QtCore.QSortFilterProxyModel):
+    def lessThan(self, x_index, y_index):
+        x_order = x_index.data(TASK_ORDER_ROLE)
+        y_order = y_index.data(TASK_ORDER_ROLE)
+        if x_order is not None and y_order is not None:
+            if x_order < y_order:
+                return True
+            if x_order > y_order:
+                return False
+
+        elif x_order is None and y_order is not None:
+            return True
+
+        elif y_order is None and x_order is not None:
+            return False
+
+        x_name = x_index.data(QtCore.Qt.DisplayRole)
+        y_name = y_index.data(QtCore.Qt.DisplayRole)
+        if x_name == y_name:
+            return True
+
+        if x_name == tuple(sorted((x_name, y_name)))[0]:
+            return True
+        return False
+
+
+class CreateWidgetTasksWidget(QtWidgets.QWidget):
+    """Widget showing active Tasks
+
+    Deprecated:
+        This widget will be removed soon. Please do not use it in new code.
+    """
+
+    task_changed = QtCore.Signal()
+
     def __init__(self, controller, parent):
         self._controller = controller
-        super(CreateWidgetTasksWidget, self).__init__(None, parent)
 
         self._enabled = None
 
-    def _create_source_model(self):
-        return TasksModel(self._controller)
+        super(CreateWidgetTasksWidget, self).__init__(parent)
+
+        tasks_view = DeselectableTreeView(self)
+        tasks_view.setIndentation(0)
+        tasks_view.setSortingEnabled(True)
+        tasks_view.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
+
+        header_view = tasks_view.header()
+        header_view.setSortIndicator(0, QtCore.Qt.AscendingOrder)
+
+        tasks_model = TasksModel(self._controller)
+        tasks_proxy = TasksProxyModel()
+        tasks_proxy.setSourceModel(tasks_model)
+        tasks_view.setModel(tasks_proxy)
+
+        layout = QtWidgets.QVBoxLayout(self)
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.addWidget(tasks_view)
+
+        selection_model = tasks_view.selectionModel()
+        selection_model.selectionChanged.connect(self._on_task_change)
+
+        self._tasks_model = tasks_model
+        self._tasks_proxy = tasks_proxy
+        self._tasks_view = tasks_view
+
+        self._last_selected_task_name = None
+
+    def refresh(self):
+        self._tasks_model.refresh()
+
+    def set_asset_id(self, asset_id):
+        # Try and preserve the last selected task and reselect it
+        # after switching assets. If there's no currently selected
+        # asset keep whatever the "last selected" was prior to it.
+        current = self.get_selected_task_name()
+        if current:
+            self._last_selected_task_name = current
+
+        self._tasks_model.set_asset_id(asset_id)
+
+        if self._last_selected_task_name:
+            self.select_task_name(self._last_selected_task_name)
+
+        # Force a task changed emit.
+        self.task_changed.emit()
+
+    def _clear_selection(self):
+        selection_model = self._tasks_view.selectionModel()
+        selection_model.clearSelection()
+
+    def select_task_name(self, task_name):
+        """Select a task by name.
+
+        If the task does not exist in the current model then selection is only
+        cleared.
+
+        Args:
+            task_name (str): Name of the task to select.
+
+        """
+        task_view_model = self._tasks_view.model()
+        if not task_view_model:
+            return
+
+        # Clear selection
+        selection_model = self._tasks_view.selectionModel()
+        selection_model.clearSelection()
+
+        # Select the task
+        mode = (
+            QtCore.QItemSelectionModel.Select
+            | QtCore.QItemSelectionModel.Rows
+        )
+        for row in range(task_view_model.rowCount()):
+            index = task_view_model.index(row, 0)
+            name = index.data(TASK_NAME_ROLE)
+            if name == task_name:
+                selection_model.select(index, mode)
+
+                # Set the currently active index
+                self._tasks_view.setCurrentIndex(index)
+                break
+
+        last_selected_task_name = self.get_selected_task_name()
+        if last_selected_task_name:
+            self._last_selected_task_name = last_selected_task_name
+
+        if not self._enabled:
+            current = self.get_selected_task_name()
+            if current:
+                self._last_selected_task_name = current
+            self._clear_selection()
+
+    def get_selected_task_name(self):
+        """Return name of task at current index (selected)
+
+        Returns:
+            str: Name of the current task.
+
+        """
+        index = self._tasks_view.currentIndex()
+        selection_model = self._tasks_view.selectionModel()
+        if index.isValid() and selection_model.isSelected(index):
+            return index.data(TASK_NAME_ROLE)
+        return None
+
+    def get_selected_task_type(self):
+        index = self._tasks_view.currentIndex()
+        selection_model = self._tasks_view.selectionModel()
+        if index.isValid() and selection_model.isSelected(index):
+            return index.data(TASK_TYPE_ROLE)
+        return None
 
     def set_asset_name(self, asset_name):
         current = self.get_selected_task_name()
@@ -163,14 +311,6 @@ class CreateWidgetTasksWidget(TasksWidget):
         # Force a task changed emit.
         self.task_changed.emit()
 
-    def select_task_name(self, task_name):
-        super(CreateWidgetTasksWidget, self).select_task_name(task_name)
-        if not self._enabled:
-            current = self.get_selected_task_name()
-            if current:
-                self._last_selected_task_name = current
-            self._clear_selection()
-
     def set_enabled(self, enabled):
         self._enabled = enabled
         if not enabled:
@@ -181,3 +321,6 @@ class CreateWidgetTasksWidget(TasksWidget):
 
         elif self._last_selected_task_name is not None:
             self.select_task_name(self._last_selected_task_name)
+
+    def _on_task_change(self):
+        self.task_changed.emit()
